Added prevention for in-container privilege escalation
Also added some qol to image creation. If the container is shutdown, the
script will now delete the container and proceed with creating a new
one. Prior, it would just throw an error since it was only looking for
containers that were still running.

+++ docker_setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [ "$(docker ps -q -f name=pastebin_container)" ]; then
+if [ "$(docker ps -a -q -f name=pastebin_container)" ]; then
     echo "Stopping and removing existing container..."
     docker stop pastebin_container
     docker rm pastebin_container
@@ -10,6 +10,6 @@ echo "Building Docker image..."
 docker build -t pastebin_app .
 
 echo "Running Docker container..."
-docker run -d -p 5000:5000 --cap-drop=ALL --cap-add=NET_BIND_SERVICE --name pastebin_container pastebin_app
+docker run -d -p 5000:5000 --security-opt=no-new-privileges --cap-drop=ALL --cap-add=NET_BIND_SERVICE --name pastebin_container pastebin_app
 
 echo "Docker container is up and running."
